more code improve

+++ flask_appbuilder/models/sqla/interface.py
@@ -833,15 +833,13 @@ class SQLAInterface(BaseInterface):
 
     def _delete_files(self, item: Model):
         for file_col in self.get_file_column_list():
-            if self.is_file(file_col):
-                if getattr(item, file_col):
-                    fm = FileManager()
-                    fm.delete_file(getattr(item, file_col))
+            if self.is_file(file_col) and getattr(item, file_col):
+                fm = FileManager()
+                fm.delete_file(getattr(item, file_col))
         for file_col in self.get_image_column_list():
-            if self.is_image(file_col):
-                if getattr(item, file_col):
-                    im = ImageManager()
-                    im.delete_file(getattr(item, file_col))
+            if self.is_image(file_col) and getattr(item, file_col):
+                im = ImageManager()
+                im.delete_file(getattr(item, file_col))
 
     """
     ------------------------------
@@ -920,11 +918,11 @@ class SQLAInterface(BaseInterface):
         """
         Returns all model's columns except pk or fk
         """
-        ret_lst = []
-        for col_name in self.get_columns_list():
-            if (not self.is_pk(col_name)) and (not self.is_fk(col_name)):
-                ret_lst.append(col_name)
-        return ret_lst
+        return [
+            col_name
+            for col_name in self.get_columns_list()
+            if (not self.is_pk(col_name)) and (not self.is_fk(col_name))
+        ]
 
     # TODO get different solution, more integrated with filters
     def get_search_columns_list(self) -> List[str]:
